feat: worked with pagination && add zod

+++ backend/src/modules/user/user.controller.ts
@@ -1,11 +1,15 @@
 import {Request, Response, NextFunction} from 'express'
 import * as userService from '../user/user.service';
-
+import {CreateUserSchema} from '../user/user.types'
 
 export const getUsers = async (req: Request, res: Response, next: NextFunction) => {
   try {
-    const users = await userService.getAllUsers();
-    res.json({success: true, data: users});
+    const page = Number(req.query.page) || 1;
+    const limit = Number(req.query.limit) || 10;
+    const search = (req.query.search as string) || "";
+
+    const {users, total} = await userService.getAllUsers(page, limit, search);
+    res.json({success: true, data: {users, total}})
   } catch (err) {
     next(err);
   }
@@ -13,22 +17,35 @@ export const getUsers = async (req: Request, res: Response, next: NextFunction) 
 
 export const createUser = async (req: Request, res: Response, next: NextFunction) => {
   try {
-    const {name, email, password} = req.body;
-    await userService.createUser({name, email, password});
-    const users = await userService.getAllUsers();
-    res.status(201).json({success: true, data: users});
+    const parsed = CreateUserSchema.parse(req.body);
+
+    const newUser = await userService.createUser(parsed);
+    res.status(201).json({ success: true, data: { newUser } });
   } catch (error) {
     next(error);
   }
 }
 
+
 export const deleteUser = async (req: Request, res: Response, next: NextFunction) => {
   try {
     const id = Number(req.params.id);
-    await userService.deleteUser(id);
-    const users = await userService.getAllUsers();
-    res.json({success: true, data: users});
+    if (isNaN(id)) {
+      return res.status(400).json({ success: false, message: "Invalid user ID" });
+    }
+
+    const deletedUser = await userService.deleteUser(id);
+
+    if (!deletedUser) {
+      return res.status(404).json({ success: false, message: "User not found" });
+    }
+
+    res.status(200).json({
+      success: true,
+      message: `User ${deletedUser.name} deleted successfully`,
+      data: deletedUser,
+    });
   } catch (error) {
     next(error);
   }
-}
+};
